✨ V15617 release: extract apply command pipeline
