+++ .agentplane/tasks/202604172036-V15617/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202604172036-V15617/release-apply-phases",
+  "created_at": "2026-04-18T05:41:55.944Z",
+  "head_sha": "a26f4be516f65a930c215f264f402d156df06383",
+  "last_verified_at": "2026-04-18T05:41:55.812Z",
+  "last_verified_sha": "a26f4be516f65a930c215f264f402d156df06383",
+  "schema_version": 1,
+  "task_id": "202604172036-V15617",
+  "updated_at": "2026-04-18T05:41:55.944Z",
+  "verify": {
+    "status": "pass"
+  }
+}
