+++ .agentplane/tasks/202604172036-V15617/pr/review.md
@@ -1,0 +1,58 @@
+# PR Review
+
+Created: 2026-04-18T05:41:55.944Z
+Branch: task/202604172036-V15617/release-apply-phases
+
+## Summary
+
+Decompose release apply command into explicit phases
+
+Split the release apply command into deterministic planning, validation, mutation, and reporting phases so the release workflow stops relying on one oversized orchestration file.
+
+## Scope
+
+- In scope: Split the release apply command into deterministic planning, validation, mutation, and reporting phases so the release workflow stops relying on one oversized orchestration file.
+- Out of scope: unrelated refactors not required for "Decompose release apply command into explicit phases".
+
+## Verification
+
+### Plan
+
+1. Run `bun run lint:core`. Expected: it succeeds and confirms the requested outcome for this task.
+2. Run `bun run test:fast`. Expected: it succeeds and confirms the requested outcome for this task.
+3. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+4. Compare the final result against the task summary and touched scope. Expected: remaining follow-up is either resolved or explicit in ## Findings.
+
+### Current Status
+
+- State: ok
+- Note: Release apply/release candidate now run through a shared pipeline module; lint:core and test:fast are green, and release apply acceptance tests stayed green after the extraction.
+
+## Risks
+
+- Risk level: not recorded
+- Breaking change: no
+
+### Rollback
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-04-18T05:41:55.944Z
+- Branch: task/202604172036-V15617/release-apply-phases
+- Head: a26f4be516f6
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
